requirements.txt
transformers
torch
python-dotenv
bleak
flask
flask-socketio
numpy
sentencepiece